requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-semantic-conventions
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
